requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
celery==5.2.7
drf-yasg==1.20.0
Django==4.0.6
djangorestframework==3.14.0
eventlet==0.33.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.4
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
self==2020.12.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.63.0
typing_extensions==4.3.0
django-rest-passwordreset
